backend/requirements.txt
fastapi[standard]
python-dotenv

SQLAlchemy
psycopg2-binary

pymysql
cryptography